Create empty frontend and backend projects
https://github.com/kjeganathan/peer-recognition-tool/issues/11

Two packages for frontend and backend, prt-frontend and prt-backend
respectively, have been added for each team to work on. The packages in
each are based on the code in the MERN tutorial [1] we followed as a
group. We're still hoping everyone can try this tutorial (it's very
short) to get a feel for all the tools.

[1] https://codingthesmartway.com/the-mern-stack-tutorial-building-a-react-crud-application-from-start-to-finish-part-1/

- Create prt-frontend package using create-react-app and Bootstrap
- Create prt-backend package using Express and Mongoose
- Add .editorconfig