Set up backend tests
https://github.com/kjeganathan/peer-recognition-tool/issues/14

Following a tutorial [1], some empty files and a dummy Jest test have
been added to the backend project. App.js serves as a foundation to
build our first few routes in.

The file contents are mostly the same as the tutorial. The call to
app.listen was moved out of app.js and into server.js so Jest can run
the tests without starting the server. A workflow for GitHub actions
was created to automatically run our Jest tests.

Finally, some NPM scripts and a README were added to make backend
development simpler.

[1] https://www.albertgao.xyz/2017/05/24/how-to-test-expressjs-with-jest-and-supertest/

- Add jest and supertest
- Create prt-backend/README.md